requirements.txt
slack-sdk
python-dotenv
slackeventsapi
flask
pymongo
dnspython
requests
gunicorn
pytz
pandas